requirements.txt
appdirs==1.4.3
asn1crypto==0.24.0
astroid==2.1.0
attrs==18.2.0
black==18.9b0
certifi==2018.11.29
cffi==1.11.5
Click==7.0
cryptography==2.4.2
future==0.17.1
idna==2.8
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
pkg-resources==0.0.0
pycparser==2.19
pylint==2.2.2
PySocks==1.6.8
python-telegram-bot==11.1.0
six==1.12.0
toml==0.10.0
typed-ast==1.1.1
wrapt==1.10.11
